chore(core): allow missing WebMvcConfigurer class in classpath
